Client controls pixels per meter.

+++ v1/Engine/include/Physics/Box.h
@@ -38,12 +38,13 @@ namespace Engine
 
 		void createFixture();
 
-		const float getCenterXInMeters() const;
-		const float getCenterYInMeters() const;
-		const float getTopLeftXInMeters() const;
-		const float getTopLeftYInMeters() const;
-		const float getWidthInMeters() const; 
-		const float getHeightInMeters() const;
+		const double getCenterXInMeters() const;
+		const double getCenterYInMeters() const;
+		const double getTopLeftXInMeters() const;
+		const double getTopLeftYInMeters() const;
+
+		const int getWidthInMeters() const; 
+		const int getHeightInMeters() const;
 
 		const int getCenterXInPixels() const;
 		const int getCenterYInPixels() const;
